+++ include/contractor/contractor.hpp
@@ -78,20 +78,19 @@ class Contractor
   private:
     ContractorConfig config;
 
-    EdgeID
-    LoadEdgeExpandedGraph(const std::string &edge_based_graph_path,
-                          util::DeallocatingVector<extractor::EdgeBasedEdge> &edge_based_edge_list,
-                          std::vector<EdgeWeight> &node_weights,
-                          const std::string &edge_segment_lookup_path,
-                          const std::string &edge_penalty_path,
-                          const std::vector<std::string> &segment_speed_path,
-                          const std::vector<std::string> &turn_penalty_path,
-                          const std::string &nodes_filename,
-                          const std::string &geometry_filename,
-                          const std::string &datasource_names_filename,
-                          const std::string &datasource_indexes_filename,
-                          const std::string &rtree_leaf_filename,
-                          const double log_edge_updates_factor);
+    EdgeID LoadEdgeExpandedGraph(const std::string &edge_based_graph_path,
+                                 std::vector<extractor::EdgeBasedEdge> &edge_based_edge_list,
+                                 std::vector<EdgeWeight> &node_weights,
+                                 const std::string &edge_segment_lookup_path,
+                                 const std::string &edge_penalty_path,
+                                 const std::vector<std::string> &segment_speed_path,
+                                 const std::vector<std::string> &turn_penalty_path,
+                                 const std::string &nodes_filename,
+                                 const std::string &geometry_filename,
+                                 const std::string &datasource_names_filename,
+                                 const std::string &datasource_indexes_filename,
+                                 const std::string &rtree_leaf_filename,
+                                 const double log_edge_updates_factor);
 };
 }
 }
